pipeline_function/requirements.txt
azure-functions
requests
psycopg2-binary
python-dotenv